% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covFactorModel.R
\name{covFactorModel}
\alias{covFactorModel}
\title{Covariance Matrix Estimation Using Factor Model}
\usage{
covFactorModel(X, type = "Stat-PCA", econ_fact = NA, K = 1,
  orthonormal = "factor", max_iter = 10, tol = 0.001, epsilon = 0,
  Psi_struct = "diag", stock_sector_info = NA)
}
\arguments{
\item{X}{xts object of dimension \eqn{T x N}, with \eqn{T} number of observations and 
\eqn{N} number of assets}

\item{type}{string object indicating the type of factor model to be used:
\itemize{\item \code{"Macro"} - macroeconomic factor model, requires user to pass \code{econ_fact}
         \item \code{"Barra"} - BARRA Industry factor model, requires user to pass \code{stock_sector_info} 
          or colnames of \code{X} to be contained in the in-built database \code{data(stock_sector_database)}
         \item \code{"Stat-PCA"} - statistical factor model, requires user to pass number of factors \code{K} (default)
         \item \code{"stat-ML"} - Maximum likelihood estimation under factor model structure }}

\item{econ_fact}{xts object of dimension \eqn{T x K}, required and used when \code{type = "Macro"}}

\item{K}{number of factors when build a statistical factor model, used when \code{type = "Stat-PCA"/"Stat-ML"} (default: \eqn{1})}

\item{orthonormal}{string object indicating position of normalization in the statistical factor 
model, used when \code{type = "Stat-PCA"}
\itemize{\item \code{"factor"} - covariance matrix of factors is identity (default)
         \item \code{"beta"} - columns of beta are orthonormal}}

\item{max_iter}{positive integer indicating maximum number of iterations, 
used when \code{type = "Stat-PCA"/"Stat-ML"} (default: \eqn{10})}

\item{tol}{double object indicating relative tolerance to determine convergence, 
used when \code{type = "Stat-PCA"/"Stat-ML"} (default: \eqn{0.001})}

\item{epsilon}{nonnegative scale indicating lower bound of residual covariance matrix's diagonal elements, 
used when \code{type = "Stat-ML"} (default: \eqn{0})}

\item{Psi_struct}{string indicating type of structure imposed on the covariance matrix of the residuals, \code{Psi},
used when \code{rtn_Sigma = TRUE}
\itemize{\item \code{"scaled_identity"} - \code{Psi} is a scale identity matrix
         \item \code{"diag"} - \code{Psi} is a diagonal matrix (default)
         \item \code{"block_diag"} - \code{Psi} is a block diagonal matrix, user required to pass \code{stock_sector_info} 
         to determine the structure of the blocks
         \item \code{"full"} - \code{Psi} is a full matrix}}

\item{stock_sector_info}{positive integer vector of length \eqn{N}, used when \code{type = "Barra"} or \code{Psi_struct = "block_diag"}}
}
\value{
matrix of dimension \eqn{N x N}, the covariance matrix
}
\description{
Estimate covariance matrix through factor model.
}
\details{
see \code{\link{factorModel}}
}
\examples{
# generate synthetic data
set.seed(234)
K <- 1   # number of factors
N <- 400  # number of stocks
mu <- rep(0, N)
beta <- mvrnorm(N, rep(1,K), diag(K)/10)
Sigma <- beta \%*\% t(beta) + diag(N)

# estimate error by loop
err_scm_vs_T <- c()
err_statPCA_diag_vs_T <- c()
index_T <- c()

for (T in N*seq(5)) {
  X <- xts(mvrnorm(T, mu, Sigma), order.by = as.Date('1995-03-15') + 1:T)
  # use statistical factor model by PCA
  cov_statPCA_diag <- covFactorModel(X, K = K, max_iter = 10)
  err_statPCA_diag_vs_T <- c(err_statPCA_diag_vs_T, norm(Sigma - cov_statPCA_diag, "F")^2)
  # use sample covariance matrix
  err_scm_vs_T <- c(err_scm_vs_T, norm(Sigma - cov(X), "F")^2)
  index_T <- c(index_T, T)
}
res <- rbind(index_T/N, err_scm_vs_T, err_statPCA_diag_vs_T)
rownames(res) <- c("T/N", "SCM", "stat-PCA + diag")
print(res)
}
\references{
K. Khamaru and R. Mazumder, “Computation of the maximum likelihood estimator in low-rank factor analysis,” 
\emph{arXiv preprint arXiv:1801.05935,} 2018.
}
\author{
ZHOU Rui & Daniel P. Palomar
}
