% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSectorInfo.R
\name{getSectorInfo}
\alias{getSectorInfo}
\title{Sector Information of Stocks}
\usage{
getSectorInfo(stocks, stock_sector_database)
}
\arguments{
\item{stocks}{vector containing the stock names}

\item{stock_sector_database}{matrix containing the pairing of stocks and sectors (first column contains the stock names and second column sector names)}
}
\value{
A list with the following components:
        \item{\code{stock_sector_info}}{integer vector with each number being an index representing one specific sector}
        \item{\code{sectors}}{string vector containing sector names sorted corresponding to their sector index}
}
\description{
Obtains the sector information of a set of stocks.
}
\examples{
mystocks <- c("AAPL",  "ABBV", "AET", "AMD", "APD", "AA","AVY", "A", "ADI", "IBM")
getSectorInfo(mystocks)
}
\author{
ZHOU Rui and Daniel P. Palomar
}
